Added File Opts and HTTP Pass/Takeover
This commit adds file options, which enable backends and frontends to store simple
strings along with the file. The intension is to enable things like NSFW blockers.

Additionally the backend is now capable of returning a special error class, the
ErrHTTPOptions. This is not an actual error and frontends that are not prepared
for this will return errors when it's passed back.

The special error allows a backend to set cookies or header and even taking over
the HTTP response completely if necessary.

This means that backends that cannot operate in a single request can show custom
pages to the user or set login information or other such stuff.

Theoretically this also enables a FCGI/CGI backend.